Move socket.io setup from App to GameContainer
